IPD: fix Ask contract placement and docs
- Point contracts to ask-report.v1.json
- Drop src/Adventorator/ask/ from plans; use commands/ for /ask handler
- Update EPIC and stories to reflect correct paths and flags
