+++ contracts/ask/v1/ask-report.v1.json
@@ -1,0 +1,37 @@
+{
+  "openapi": "3.1.0",
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "ask.ask_report.v1",
+  "description": "Contract for AskReport emitted by /ask pipeline (runtime models live in src/Adventorator/schemas.py).",
+  "type": "object",
+  "properties": {
+    "version": {"type": "string", "const": "1.0"},
+    "raw_text": {"type": "string"},
+    "intent": {
+      "type": "object",
+      "properties": {
+        "action": {"type": "string"},
+        "actor_ref": {"type": ["string", "null"]},
+        "target_ref": {"type": ["string", "null"]},
+        "modifiers": {"type": "array", "items": {"type": "string"}}
+      },
+      "required": ["action"],
+      "additionalProperties": false
+    },
+    "tags": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "properties": {
+          "key": {"type": "string"},
+          "value": {"type": ["string", "null"]},
+          "confidence": {"type": "number", "minimum": 0.0, "maximum": 1.0}
+        },
+        "required": ["key"],
+        "additionalProperties": false
+      }
+    }
+  },
+  "required": ["version", "raw_text", "intent", "tags"],
+  "additionalProperties": false
+}
